+++ RELEASES.md
@@ -7,6 +7,9 @@ Released on TBD (UTC).
 - Added `subscribe_funding_rates(...)` and `unsubscribe_funding_rates(...)` methods for actors
 - Added `on_funding_rate(...)` handler for actors
 - Added `funding_rate(...)` and `add_funding_rate(...)` for `Cache`
+- Added `due_post_only` field for `OrderRejected` event, only properly populated for Binance and Bybit for now
+- Added `log_rejected_due_post_only_as_warning` config option for `StrategyConfig` (default `True` to retain current behavior)
+- Added `log_rejected_due_post_only_as_warning` config option for `BinanceExecClientConfig` (default `True` to retain current behavior)
 - Added support for additional Databento schemas: `OHLCV_EOD`, `CMBP_1`, `CBBO_1S`, `CBBO_1M`, and `TCBBO`
 - Added configurable schema parameters for Databento quote and trade subscriptions, allowing `TBBO`/`TCBBO` for efficient combined data feeds
 - Added support for option combos for Interactive Brokers (#2812), thanks @faysou